Revert "Configurable caching of blocking threads"
