+++ core/jvm/src/main/scala/cats/effect/unsafe/WorkStealingThreadPool.scala
@@ -34,7 +34,6 @@ import cats.effect.tracing.TracingConstants
 
 import scala.collection.mutable
 import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.Duration
 
 import java.util.Comparator
 import java.util.concurrent.{ConcurrentSkipListSet, ThreadLocalRandom}
@@ -100,8 +99,6 @@ private[effect] final class WorkStealingThreadPool(
 
   private[unsafe] val cachedThreads: ConcurrentSkipListSet[WorkerThread] =
     new ConcurrentSkipListSet(Comparator.comparingInt[WorkerThread](_.nameIndex))
-  private[unsafe] def runtimeBlockingExpiration: Duration =
-    self.config.runtimeBlockingExpiration
 
   /**
    * The shutdown latch of the work stealing thread pool.
